4to commit

+++ app/src/main/java/com/example/movies2/data/network/MovieApiClient.kt
@@ -9,8 +9,8 @@ import retrofit2.http.Path
 interface MovieApiClient {
 
     @GET("movie/{movie_id}")
-    suspend fun getMovieById(@Path("movie_id") id:Long): Response<MovieModel?>
+    suspend fun getMovieById(@Path("movie_id") id:Int): Response<MovieModel?>
 
     @GET("list/{list_id}")
-    suspend fun getMoviesList(@Path("list_id") id:Long) : Response<MoviesListModel?>
+    suspend fun getMoviesList(@Path("list_id") id:Int) : Response<MoviesListModel?>
 }
